+++ addon-manager/api/pom.xml
@@ -32,6 +32,18 @@
             </exclusion>
          </exclusions>
       </dependency>
+      <dependency>
+         <groupId>org.jboss.forge.addon</groupId>
+         <artifactId>javaee-spi</artifactId>
+         <classifier>forge-addon</classifier>
+         <scope>provided</scope>
+      </dependency>
+      <dependency>
+         <groupId>org.jboss.forge.addon</groupId>
+         <artifactId>configuration</artifactId>
+         <classifier>forge-addon</classifier>
+         <scope>provided</scope>
+      </dependency>
 
       <dependency>
          <groupId>org.jboss.forge.addon</groupId>
