+++ addon-manager/api/src/main/java/org/jboss/forge/addon/manager/catalog/AddonDescriptorCatalog.java
@@ -5,7 +5,7 @@
  * http://www.eclipse.org/legal/epl-v10.html
  */
 
-package org.jboss.forge.addon.manager.impl.catalog;
+package org.jboss.forge.addon.manager.catalog;
 
 import java.util.List;
