+++ addon-manager/impl/pom.xml
@@ -34,12 +34,12 @@
          <classifier>forge-addon</classifier>
          <scope>provided</scope>
       </dependency>
-      <dependency>
+      <!-- <dependency>
          <groupId>org.jboss.forge.addon</groupId>
          <artifactId>javaee-spi</artifactId>
          <classifier>forge-addon</classifier>
          <scope>provided</scope>
-      </dependency>
+      </dependency> -->
       <dependency>
          <groupId>org.jboss.forge.addon</groupId>
          <artifactId>convert</artifactId>
@@ -94,12 +94,6 @@
       <dependency>
          <groupId>org.jboss.forge.addon</groupId>
          <artifactId>projects</artifactId>
-         <classifier>forge-addon</classifier>
-         <scope>provided</scope>
-      </dependency>
-      <dependency>
-         <groupId>org.jboss.forge.addon</groupId>
-         <artifactId>configuration</artifactId>
          <classifier>forge-addon</classifier>
          <scope>provided</scope>
       </dependency>
