+++ addon-manager/impl/src/main/java/org/jboss/forge/addon/manager/impl/ui/catalog/AddonSearchCommand.java
@@ -10,8 +10,8 @@ package org.jboss.forge.addon.manager.impl.ui.catalog;
 import java.io.PrintStream;
 import java.util.List;
 
-import org.jboss.forge.addon.manager.impl.catalog.AddonDescriptor;
-import org.jboss.forge.addon.manager.impl.catalog.AddonDescriptorCatalogRegistry;
+import org.jboss.forge.addon.manager.catalog.AddonDescriptor;
+import org.jboss.forge.addon.manager.catalog.AddonDescriptorCatalogRegistry;
 import org.jboss.forge.addon.manager.impl.ui.AddonCommandConstants;
 import org.jboss.forge.addon.manager.impl.ui.AddonInstallCommand;
 import org.jboss.forge.addon.ui.command.UICommand;
